refactor: Move get_users_with_role to utils.user
